+++ tests/dummy/app/Policies/UserPolicy.php
@@ -1,0 +1,27 @@
+<?php
+
+namespace DummyApp\Policies;
+
+use DummyApp\User;
+
+class UserPolicy
+{
+
+    /**
+     * @param User $user
+     * @return bool
+     */
+    public function author(User $user)
+    {
+        return $user->author;
+    }
+
+    /**
+     * @param User $user
+     * @return bool
+     */
+    public function admin(User $user)
+    {
+        return $user->admin;
+    }
+}