+++ tests/dummy/app/User.php
@@ -35,6 +35,8 @@ class User extends Authenticatable
         'name',
         'email',
         'password',
+        'author',
+        'admin',
     ];
 
     /**
@@ -45,6 +47,22 @@ class User extends Authenticatable
     protected $hidden = [
         'password',
         'remember_token',
+    ];
+
+    /**
+     * @var array
+     */
+    protected $casts = [
+        'author' => 'bool',
+        'admin' => 'bool',
+    ];
+
+    /**
+     * @var array
+     */
+    protected $attributes = [
+        'author' => false,
+        'admin' => false,
     ];
 
     /**