+++ tests/dummy/database/factories/ModelFactory.php
@@ -78,8 +78,14 @@ $factory->define(DummyApp\Post::class, function (Faker $faker) {
         'slug' => $faker->unique()->slug,
         'content' => $faker->text,
         'author_id' => function () {
-            return factory(DummyApp\User::class)->create()->getKey();
+            return factory(DummyApp\User::class)->states('author')->create()->getKey();
         },
+    ];
+});
+
+$factory->state(DummyApp\Post::class, 'published', function (Faker $faker) {
+    return [
+        'published_at' => $faker->dateTimeBetween('-1 month', 'now'),
     ];
 });
 
@@ -97,6 +103,14 @@ $factory->define(DummyApp\User::class, function (Faker $faker) {
         'email' => $faker->unique()->email,
         'password' => bcrypt(str_random(10)),
     ];
+});
+
+$factory->state(DummyApp\User::class, 'author', function () {
+    return ['author' => true];
+});
+
+$factory->state(DummyApp\User::class, 'admin', function () {
+    return ['admin' => true];
 });
 
 /** Video */