+++ tests/dummy/database/migrations/2018_02_11_1648_create_tables.php
@@ -32,6 +32,8 @@ class CreateTables extends Migration
             $table->string('name');
             $table->string('email')->unique();
             $table->string('password');
+            $table->boolean('author');
+            $table->boolean('admin');
             $table->rememberToken();
             $table->timestamps();
             $table->unsignedInteger('country_id')->nullable();